Proof of concept refactor of navigation lazy-load scrolling for reuse.
