+++ assets/js/components/DashboardNavigation/Navigation.js
@@ -27,7 +27,7 @@ import { Chip } from '@material/react-chips';
 /**
  * WordPress dependencies
  */
-import { useState, useEffect, useCallback, useRef } from '@wordpress/element';
+import { useState, useEffect, useRef } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 
 /**
@@ -46,33 +46,24 @@ import {
 	ANCHOR_ID_KEY_METRICS,
 	ANCHOR_ID_TRAFFIC,
 } from '../../googlesitekit/constants';
-import { CORE_USER } from '../../googlesitekit/datastore/user/constants';
-import { CORE_WIDGETS } from '../../googlesitekit/widgets/datastore/constants';
 import {
 	CORE_UI,
 	ACTIVE_CONTEXT_ID,
 } from '../../googlesitekit/datastore/ui/constants';
-import {
-	CONTEXT_ENTITY_DASHBOARD_TRAFFIC,
-	CONTEXT_ENTITY_DASHBOARD_CONTENT,
-	CONTEXT_ENTITY_DASHBOARD_SPEED,
-	CONTEXT_ENTITY_DASHBOARD_MONETIZATION,
-	CONTEXT_MAIN_DASHBOARD_KEY_METRICS,
-	CONTEXT_MAIN_DASHBOARD_TRAFFIC,
-	CONTEXT_MAIN_DASHBOARD_CONTENT,
-	CONTEXT_MAIN_DASHBOARD_SPEED,
-	CONTEXT_MAIN_DASHBOARD_MONETIZATION,
-} from '../../googlesitekit/widgets/default-contexts';
-import useDashboardType, {
-	DASHBOARD_TYPE_MAIN,
-} from '../../hooks/useDashboardType';
+import useDashboardType from '../../hooks/useDashboardType';
 import { useBreakpoint } from '../../hooks/useBreakpoint';
 import { getContextScrollTop } from '../../util/scroll';
 import { trackEvent } from '../../util';
 import useViewContext from '../../hooks/useViewContext';
-import useViewOnly from '../../hooks/useViewOnly';
-import { useFeature } from '../../hooks/useFeature';
-const { useSelect, useDispatch } = Data;
+import {
+	useGetDefaultChipID,
+	useIsJumpingTo,
+	useScrollToContextArea,
+	useShowContextAreas,
+	useShowKeyMetrics,
+} from './DashboardNavigationContext';
+import { useCallback } from 'react';
+const { useDispatch } = Data;
 
 export default function Navigation() {
 	const dashboardType = useDashboardType();
@@ -81,119 +72,19 @@ export default function Navigation() {
 
 	const initialHash = global.location.hash?.substring( 1 );
 	const [ selectedID, setSelectedID ] = useState( initialHash );
-	const [ isJumpingTo, setIsJumpingTo ] = useState(
-		initialHash || undefined
-	);
+	const [ isJumpingTo, setIsJumpingTo ] = useIsJumpingTo();
 	const [ isSticky, setIsSticky ] = useState( false );
 
 	const viewContext = useViewContext();
-	const viewOnlyDashboard = useViewOnly();
 
 	const { setValue } = useDispatch( CORE_UI );
 
-	const viewableModules = useSelect( ( select ) => {
-		if ( ! viewOnlyDashboard ) {
-			return null;
-		}
+	const showKeyMetrics = useShowKeyMetrics();
 
-		return select( CORE_USER ).getViewableModules();
-	} );
+	const { showTraffic, showContent, showSpeed, showMonetization } =
+		useShowContextAreas();
 
-	const keyMetricsEnabled = useFeature( 'keyMetrics' );
-
-	const isKeyMetricsWidgetHidden = useSelect(
-		( select ) =>
-			keyMetricsEnabled && select( CORE_USER ).isKeyMetricsWidgetHidden()
-	);
-
-	const widgetContextOptions = {
-		modules: viewableModules ? viewableModules : undefined,
-	};
-
-	const showKeyMetrics = useSelect( ( select ) => {
-		if (
-			! keyMetricsEnabled ||
-			dashboardType !== DASHBOARD_TYPE_MAIN ||
-			isKeyMetricsWidgetHidden === true
-		) {
-			return false;
-		}
-
-		return select( CORE_WIDGETS ).isWidgetContextActive(
-			CONTEXT_MAIN_DASHBOARD_KEY_METRICS,
-			widgetContextOptions
-		);
-	} );
-
-	const showTraffic = useSelect( ( select ) =>
-		select( CORE_WIDGETS ).isWidgetContextActive(
-			dashboardType === DASHBOARD_TYPE_MAIN
-				? CONTEXT_MAIN_DASHBOARD_TRAFFIC
-				: CONTEXT_ENTITY_DASHBOARD_TRAFFIC,
-			widgetContextOptions
-		)
-	);
-
-	const showContent = useSelect( ( select ) =>
-		select( CORE_WIDGETS ).isWidgetContextActive(
-			dashboardType === DASHBOARD_TYPE_MAIN
-				? CONTEXT_MAIN_DASHBOARD_CONTENT
-				: CONTEXT_ENTITY_DASHBOARD_CONTENT,
-			widgetContextOptions
-		)
-	);
-
-	const showSpeed = useSelect( ( select ) =>
-		select( CORE_WIDGETS ).isWidgetContextActive(
-			dashboardType === DASHBOARD_TYPE_MAIN
-				? CONTEXT_MAIN_DASHBOARD_SPEED
-				: CONTEXT_ENTITY_DASHBOARD_SPEED,
-			widgetContextOptions
-		)
-	);
-
-	const showMonetization = useSelect( ( select ) =>
-		select( CORE_WIDGETS ).isWidgetContextActive(
-			dashboardType === DASHBOARD_TYPE_MAIN
-				? CONTEXT_MAIN_DASHBOARD_MONETIZATION
-				: CONTEXT_ENTITY_DASHBOARD_MONETIZATION,
-			widgetContextOptions
-		)
-	);
-	const getDefaultChipID = useCallback( () => {
-		if ( showKeyMetrics ) {
-			return ANCHOR_ID_KEY_METRICS;
-		}
-
-		if ( ! viewOnlyDashboard ) {
-			return ANCHOR_ID_TRAFFIC;
-		}
-
-		if ( showTraffic ) {
-			return ANCHOR_ID_TRAFFIC;
-		}
-
-		if ( showContent ) {
-			return ANCHOR_ID_CONTENT;
-		}
-
-		if ( showSpeed ) {
-			return ANCHOR_ID_SPEED;
-		}
-
-		if ( showMonetization ) {
-			return ANCHOR_ID_MONETIZATION;
-		}
-
-		return '';
-	}, [
-		showKeyMetrics,
-		showTraffic,
-		showContent,
-		showSpeed,
-		showMonetization,
-		viewOnlyDashboard,
-	] );
+	const getDefaultChipID = useGetDefaultChipID();
 
 	const isValidChipID = ( chipID ) => {
 		if ( showKeyMetrics && chipID === ANCHOR_ID_KEY_METRICS ) {
@@ -219,29 +110,16 @@ export default function Navigation() {
 		return false;
 	};
 
+	const scrollToContextArea = useScrollToContextArea();
+
 	const handleSelect = useCallback(
 		( { target } ) => {
 			const chip = target.closest( '.mdc-chip' );
 			const chipID = chip?.dataset?.contextId; // eslint-disable-line sitekit/acronym-case
 
-			global.history.replaceState( {}, '', `#${ chipID }` );
-
-			setIsJumpingTo( chipID );
-			trackEvent( `${ viewContext }_navigation`, 'tab_select', chipID );
-
-			global.scrollTo( {
-				top:
-					chipID !== getDefaultChipID()
-						? getContextScrollTop( `#${ chipID }`, breakpoint )
-						: 0,
-				behavior: 'smooth',
-			} );
-
-			setTimeout( () => {
-				setValue( ACTIVE_CONTEXT_ID, chipID );
-			}, 50 );
+			scrollToContextArea( chipID );
 		},
-		[ breakpoint, viewContext, setValue, getDefaultChipID ]
+		[ scrollToContextArea ]
 	);
 
 	useMount( () => {
@@ -372,6 +250,7 @@ export default function Navigation() {
 		viewContext,
 		setValue,
 		getDefaultChipID,
+		setIsJumpingTo,
 	] );
 
 	return (
